server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-multipart==0.0.9
boto3==1.34.34
google-cloud-storage==2.14.0
fpdf2==2.7.8
alembic==1.13.1
openai==1.75.0
anthropic==0.40.0
firebase-admin==6.5.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
google-auth==2.29.0
python-dotenv==1.0.1
google-generativeai==0.8.0
pymupdf==1.24.0
httpx==0.27.0
python-dateutil==2.9.0
stripe==8.0.0
slowapi==0.1.9
python-magic==0.4.27
email-validator==2.1.1
